working on restricting access if they do not have an account. Have a better understanding about the routes of this program now
