+++ client/src/components/Nav/index.js
@@ -1,28 +1,91 @@
-import React from "react";
-import "./style.css";
+import React, { Component } from 'react'
+import { Redirect } from 'react-router-dom'
+import { Route, Link } from 'react-router-dom'
+import './style.css';
+import axios from 'axios'
 
-function Nav() {
-  return (
-    <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
-      <a className="navbar-brand" href="/">
-      Bit-Pulse <i class="fab fa-bitcoin"></i>
-      </a>
+class Navbar extends Component {
+    constructor() {
+        super()
+        this.logout = this.logout.bind(this)
+    }
 
-      <div id="navbarNav">
-                <ul className="navbar-nav">
+    logout(event) {
+        event.preventDefault()
+        console.log('logging out')
+        axios.post('/user/logout').then(response => {
+          console.log(response.data)
+          if (response.status === 200) {
+            this.props.updateUser({
+              loggedIn: false,
+              username: null
+            })
+          }
+          window.location.replace("/login");
+        }).catch(error => {
+            console.log('Logout error')
+        })
+      }
 
-                    <li className="nav-item" id="home">
-                        <a className="nav-link" href="/login"><button type="button" className="btn btn-dark ">Login</button></a>
-                    </li>
+    render() {
+        const loggedIn = this.props.loggedIn;
+        console.log('navbar render, props: ')
+        console.log(this.props);
+        
+        return (
+            <div>
+                {/* checks if the user is logged in */}
+                {/* displays a different navbar */}
+                {loggedIn ? ( 
+                    <header className="navbar App-header" id="nav-container"> 
+                        <div className="col-3">
+                            <section className="navbar-section">
+                                <Link to="/dashboard" className="btn btn-link text-secondary">
+                                <span className="text-secondary">Dashboard</span></Link>
+                            </section>    
+                        </div>
+                        <div className="col-6">
+                            <div id="top-filler"></div>
+                            <a className="navbar-brand" href="/">
+                            Bit-Pulse <i class="fab fa-bitcoin"></i>
+                            </a>
+                        </div>
+                        <div className="col-3" >        
+                            <section className="navbar-section">
+                                <Link to="#" className="btn btn-link text-secondary" onClick={this.logout}>
+                                <span className="text-secondary">logout</span></Link>
+                            </section>                        
+                        </div>        
+                    </header>
+                ) : (
+                <header className="navbar App-header" id="nav-container"> 
+                    <div className="col-3">
+                        <section className="navbar-section">
+                            <Link to="/" className="btn btn-link text-secondary">
+                                <span className="text-secondary">home</span>
+                                </Link>
+                            <Link to="/login" className="btn btn-link text-secondary">
+                            <span className="text-secondary">login</span>
+                                </Link>
+                            <Link to="/signup" className="btn btn-link">
+                            <span className="text-secondary">sign up</span>
+                                </Link>
+                        </section>
+                    </div>
+                    <div className="col-6">
+                        <div id="top-filler"></div>
+                        <a className="navbar-brand" href="/">
+                        Bit-Pulse <i class="fab fa-bitcoin"></i>
+                        </a>
+                    </div>
+                    <div className="col-3" > 
 
-                    <li className="nav-item" id="s">
-                        <a className="nav-link" href="/signup"><button type="button" className="btn btn-dark">Sign-Up</button></a>
-                    </li>
-                </ul>
+                    </div>
+                </header>
+                )}
             </div>
-
-    </nav>
-  );
+        )
+    }
 }
 
-export default Nav;
+export default Navbar
