--- client/src/components/home.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react'
-
-class Home extends Component {
-    constructor() {
-        super()
-    }
-
-
-    render() {
-        const imageStyle = {
-            width: 400
-        }
-        return (
-            <div>
-                <p>It's good to be home</p>
-                <img style={imageStyle} src="https://i.ytimg.com/vi/N1icEHtgb3g/maxresdefault.jpg" />
-            </div>
-        )
-
-    }
-}
-
-export default Home
